requirements.txt
# CCAP Analyzer - Dépendances Python
# Lexigency - 2026

# Framework Web
flask>=3.0
gunicorn>=21.2
werkzeug>=3.0

# Traitement Documents Word
python-docx>=1.1
lxml>=5.0

# Requêtes HTTP (OpenRouter API)
requests>=2.31

# Configuration
python-dotenv>=1.0

# Base vectorielle RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0
